feat: added Minimum Viable Product example
a short script that shows it's possible to store URLs, and retrieve them
upon request, returning them as redirects

run with:

uvicorn mw_url_shortener/mvp:app

I'm not sure if this properly runs the database query in an async
context, or if that blocks the main thread.

If I want to use PonyORM, I'll probably have to find some way of running
it in an async context.

Alternatively, I could switch to a different ORM, probably any of the
ones mentioned in the FastAPI docs, and use the techniques described
there to gain async access to the database.