refactor: reorganized project structure
server.py was broken up into:

database/
  models.py
  entities.py
  interface.py
server/
  __main__.py
  api/
    authentication.py
    __main__.py
    redirects.py
    users.py

The pydantic models will be in:
database/models.py

The pony database entities, or actual "Models" in the "Object-Relational
Model" sense, will be in:
database/entities.py

The functions that manipulate the database will be in:
database/interface.py

The main "run the server" function will be in:
server/__main__.py

The api endpoint definitions are under:
server/api/

The endpoint classes will all have their own submodules:
users -> users.py
redirects -> redirects.py
logs -> logs.py

They will be included in the API in the main api file:
server/api/__main__.py

All of the authentication will be under:
server/api/authentication.py

My biggest concern with this will be import loops. It's happened before,
where I've imported a file, that itself imports a file, that eventually
tries to import the file I'm importing from.

I read through the description of how Pony's db_session works:
https://docs.ponyorm.org/transactions.html

and I think I can get away with using them in functions that are used
with Depends(). As long as I don't call them directly from the
asynchronous path operation functions, I should be fine.

Currently, the one stubbed out API operation does call it directly, and
needs to be changed to call it using Depends().

I think I'll use pydantic's BaseSettings model for managing the
settings:

I want the main way to provide configuration to this app to be through
command line arguments. Having a file would be nice, though, and
BaseSettings has a way of including both the environment variables and
a .env file.

I'm hoping to be able to serialize the settings so they can be stored in
the database, that way, once configuration is done, the only flag that's
needed is --database-file:

mw_url_shortener server --database-file /path/to/url_shortener.sqlitedb

I still would want all configuration to be able to be taken from the
environment, .env file, and command line, but having it in the database
would be nice.