refactor: reorganize module dependencies
I wanted to make the Database() object easier to access, and I know see
that I should have taken a different approach.

I should rely on the starlette state management that FastAPI exposes:
https://www.starlette.io/applications/#storing-state-on-the-app-instance

I wanted to avoid having to connect everything outside the app, or have
a "context" object that everything accepts, but it makes it MUCH easier
to test a function if it doesn't access state from it's own environment.

Right now, I greatly dislike the interface for setting the database
file.