feat(console): add --root-path to cli
Versioned the api variables, so that if the api is ever changed
completely, and it's important to me to implement a deprecation period,
I can simply add the new api, in addition to the old one, without much
fuss. This may be preemptive refactoring.

I'm going back and forth on how to enforce restrictions on the Key type,
and a lot of the other types. I'll probably be very lax, build the
application, and note how things break in # NOTE:BUG comments, and then
settle on how to fix them.

The server run function is now in the console app, in part because the
api_key setting appears to need to be set when the sub-app is mounted on
the main server app, and in part because I'm not certain that setting
the attributes on the FastAPI object will result in the same behaviour
as passing it arguments when instantiating it.

I'm still not exactly sure where to keep this server running function,
or the settings models.

Next:

Move the api app object into an api.main.py file.

Add the following command line options to the server sub-command:
--disable-api
--docs-prefix
--disable-docs

Add a Configuration model, and store the configuration in the database
on startup, so that settings can be read from the database, simplifying
running a configured app to:
mw_url_shortener --database-file ./url.sqlitedb server

This would be a strong contender with:
mw_url_shortener --env-file ./prod.env server