refactor: too tangled
Changed the way the env_file is specified, so that the command line
option doesn't override the default value, if one is set in the class.

It looks like apistar would be a good way to programmatically give a
client that can be used from the command line.

If I want to make sure that only an already set up database is being
given, I need to make sure to test the database schema, and catch errors
so that they can be presented nicely to the user.
I would also need to ensure that the setup process only checks and does
not populate the database, as it's currently doing.

It turns out I can't simply specify NewTypes with Pony's methods. It
doesn't make sense to either, since the database will deal with base
python types. The Pydantic models are what will ensure conversion to
proper types.

The database interface, as-is, will be pretty challenging to test, and
I'm still not convinced that it works the way I want it to: using a
single, configured Database() object in all the uvicorn worker
processes.

Also, now, even the main app is a router that's wired into the main app
during startup. That's a bit much.

In general, things are getting a bit tangled. I think taking a step back
and setting up testing would be very helpful. It will be painfully slow,
but also switching to sqlalchemy and copying the fastapi template
generator's testing implementation would probably make my life
significantly easier.