feat(config): add functions to save and retrieve config
Added speedup for json encoding and decoding to all pydantic models,
following the recipe given in:
https://pydantic-docs.helpmanual.io/usage/exporting_models/#custom-json-deserialisation

Removed func from CommonSettings model. This is a breaking change, as
console.py hasn't been redone.

Added tests for the feature for a feature for the first time.

The tests are very rudimentary right now, and are run by simply running
the command:
pytest

Moved Database() Entity registration to a function with Database()
creation so that dependency injection is easier. This also means that
unless setup_db is called for EACH database object returned, the
returned Database() object can't be used.

Also, this is now back to a chicken and egg problem of where to get
settings from during runtime: they can be loaded by calling the settings
object, but that may not be populated with the values that were
available a runtime. The config will be saved to the database on
application startup, so they can be loaded from there, but the database
file path is one of the dynamic values that currently is only available
at runtime from the configuration stored in the database.

Still not sure how to solve this. Currently, I'm thinking of hardcoding
the database path, so that I can build the rest of the application, and
continue, while working on this problem on the side.