feat(database.redirect): flesh out tests and interface
There are a few tests that I probably should write, to test codepaths
like "does update_redirect allow changing the key", and other questions.

I think it'd also be a good idea to fix up the docstrings in the
utilities test file, to match the style I've adopted elsewhere, of
asking a question.

I'm also fairly certain the user and redirect tests can share more
between them: the redirect tests have more assertions, and the users
tests don't have as many different situations.

I am happy with how I wrote the user implementations, though, as they
are very easy to read. I wrote out a few duplicate ones for the
redirects, and they weren't as clear as the user ones.

I think it would be helpful, though, to add a bit more explicit casting
in those implementations, so that if a Uri ever becomes a more involved
type, as long as it has a __str__ method, it can still be passed to the
interface.

Added timeouts for the tests to ensure no key generation goes into an
endless loop. While I can watch for it on my machine, I certainly don't
want it doing that in CI.

There's also gotta be a faster or better way of testing if a Primary Key
is taken in a database or not, than to use the ORM and try to select
something by primary key.

I'm also 80% sure I could be writing a bit less code in the tests by
using FastAPI's testing dependency injection.