feat(config): starts adding config functions
I kind of went off the path here. I realized that I have 3 different
applications all bundled into one, and that I may not even want a
database file configured if I'm only using the client to interact with
the API.

In this case, I think going back to my original idea of having a base
class for the application-wide settings, and then other settings classes
can derive from that so that they can indicate which settings they'll
need. The tricky part will be making sure that all those settings will
be reconstructed in a new process.

It looks like the only way to do this, especially considering the I
can't use a database all the time to synchronize and share state, is to
set the environment to hold that data. That way, even if the original
arguments passed to the program at startup are no longer available, the
function that set the initial configuration also setup the environment
so that the name of the settings class can be retrieved, and then that
class can use the environment settings to magically pull in and
transform all of the data needed.

Also moved a lot of the first-time setup stuff to
mw_url_shortener/setup.py

Also updated some dependencies.

Because a lot of the parts of the program will soon depend on
environment variabled being set, there's now a fixture that runs during
all the tests that makes any environment variables with the prefix this
application uses invisible to the tests.