feat(config): begin adding config module functions
After the config module can get and set configuration in the
environment, dotenv file, and database, I want to redo the database
setup:

I think moving the get_db() and setup_db() to a new file would be really
helpful. I'm thinking database/setup.py

I have some misgivings about having files named setup.py that have
nothing to do with packaging, but it's the name that makes the most
sense to me.

I want to rename the get_db() function to make_db(), and have the flow
look like:

- application started
- set configuration with config.save()

Then, while the application is running, it can use get_db() to go from
having a configuration in the environment, to a database object:

- config.get()
- uses the database_file
- checks to make sure the database is already configured and set up
- builds a Database object using the file

That way, the dependency of get_db() doesn't require arguments, and can
always return a Database object, as long as the configuration is already
present, either in the module cache or the environment.

I think having a module Database object cache would also be a way to
ensure the function can return quickly.

After get_db() works, there's quite a few things left:

- write server tests
- write console tests
- add local database cli interface
- add remote api client
- add logging (with tests)