chore(database/start): correct types
Only the AsyncEngine has a .dispose() method, which is used to close all
connections in a pool. This is handy for using with multiple event
loops:
https://docs.sqlalchemy.org/en/20/orm/extensions/asyncio.html#using-multiple-asyncio-event-loops

I don't think I'll be doing that: even at the terminal, with Typer, the
app can be organized to only use one event loop.