refactor(interfaces): satisfy mypy
UserInterface is a Protocol, so uses structural typing

Resource was redone to make mypy happy. There's a few typing.cast()s in
mw_url_shortener.interfaces.helpers.get_resource() that are lying to
mypy, but I feel okay about that since they're pretty truthy lies.

This might be the one point in time where the types are okay in both the
module and the tests, and all the tests are passing.