feat(server): cleanup for testing
It looks like the OAuth2 endpoint has to be assigned before the object
is passed as the default value to a function inside a Depends(), and
then that function has to be associated with a route.

This makes passing this in in-place, even from an environment variable,
challenging.

I'll probably use a make_app() function that returns a FastAPI() object
that's had the appropriate settings applied.

This'll be part of the /cli.py startup process, and will be great to
flesh out during testing, fill out the API, then figure out how to
programmatically start the server, with this object.

That way if anything has to be refactored, there are tests in place to
help.