refactor(server/cli): make debug command more robust
I don't want to have to change too many things when hypercorn updates,
and the previous version relied on a lot of internals of hypercorn.

This way, it's fairly easy to use hypercorn, writing a start command
will be much easier, and hypercorn can be swapped out with uvicorn or
run with gunicorn very easily.