feat(utils): add combinatorics for short_links
I wanted to know, given a number of unique characters, if it takes a
certain amount of time to generate a sequence of characters chosen from
the unique set, how many sequences need to be generated before the next
one has a 70% chance of colliding with a previously generated one?

Essentially, how long should the default character set for short links,
and the default short link length be, to optimize for a low chance of
collision?

Calculating when the total space is 70% full isn't too hard:
0.7 == (number of sequences) / ((number of unique characters) ^ (short link length))

sagemath seems like a good bet for plotting out the general solution
space.