fix(cli/remote): close httpx.AsyncClient only once
httpx.AsyncClient is not re-openable: once it's been opened once in a
`with` block, and that block exits, it the whole connection pool and
can't be reopened.

There are a few places where multiple httpx.AsyncClient calls are made
in the same command (if they're in separate commands, they'll be run in
speratate invocations of the cli). Plus it'd be nice not to have to
worry about it, since the `with` blocks are internal to the interface,
and with the `run_sync` architecture, it'd have to do both calls in the
same interface function, which would require a lot of duplication of
code.

Fortunately, httpx.AsyncClient is okay being opened once, "manually",
and Click includes a builtin function for calling closers when the
script exits, even on errors, so this was a simple change.

If this interface is used elsewhere, that'll have to be kept in mind.